Reuse the "get user's admin" method by creating a Mixin
Mixins allow us to reuse code across different views.

We look up the admin for a user once in /profiles and once in /backup_codes.
